Offer_Letter_Generator/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-docx>=1.0.0
